pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "elio-cli"
version = "0.2.5"
description = "Unified AI Coding Agent — Claude, Gemini, ChatGPT in one terminal"
requires-python = ">=3.11"
dependencies = [
    "typer",
    "rich",
    "prompt-toolkit",
    "anthropic",
    "openai",
    "groq",
    "google-genai",
    "google-auth",
    "google-auth-oauthlib",
    "keyring",
    "python-dotenv",
    "aiofiles",
    "pydantic",
    "toml",
    "httpx",
]

[project.scripts]
elio = "cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["cli", "auth", "config", "providers", "session", "files", "utils"]